+++ rocprim/include/rocprim/functional.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2023 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2017-2026 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -31,23 +31,10 @@ BEGIN_ROCPRIM_NAMESPACE
 /// \addtogroup utilsmodule_functional
 /// @{
 
-/// \brief Prints the supplied error message only once (using only one of the active threads).
-/// \note Currently, this is not defined for Navi devices.
-#if ROCPRIM_NAVI
-ROCPRIM_PRAGMA_MESSAGE("GPU printf warnings for invalid rocPRIM warp operations on Navi GPUs "
-                       "temporarily disabled, due to performance issues with printf.")
-    #define ROCPRIM_PRINT_ERROR_ONCE(message) \
-        {}
-#else
-    #define ROCPRIM_PRINT_ERROR_ONCE(message)                           \
-        {                                                               \
-            unsigned int idx = threadIdx.x + (blockIdx.x * blockDim.x); \
-            idx += threadIdx.y + (blockIdx.y * blockDim.y);             \
-            idx += threadIdx.z + (blockIdx.z * blockDim.z);             \
-            if(idx == 0)                                                \
-                printf("%s\n", #message);                               \
-        }
-#endif
+/// \brief Unused macro.
+/// \deprecated Will be removed in ROCm 9.0
+#define ROCPRIM_PRINT_ERROR_ONCE(message) \
+    {}
 
 /// \brief Returns the maximum of its arguments.
 template<class T>