+++ rocprim/include/rocprim/warp/warp_reduce.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2025 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2017-2026 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -194,10 +194,7 @@ public:
         {
             if(VirtualWaveSize > ::rocprim::arch::wavefront::size())
             {
-                ROCPRIM_PRINT_ERROR_ONCE(
-                    "Specified warp size exceeds current hardware supported warp "
-                    "size. Aborting warp sort.");
-                return;
+                __builtin_trap(); // behavior undefined if virtual wave size exceeds hardware limit
             }
         }
         base_type::reduce(input, output, storage, reduce_op);
@@ -212,9 +209,7 @@ public:
         typename std::enable_if<(FunctionWarpSize > arch::wavefront::max_size()), void>::type
     {
         (void)reduce_op;
-        ROCPRIM_PRINT_ERROR_ONCE("Specified warp size exceeds current hardware supported warp "
-                                 "size. Aborting warp sort.");
-        return;
+        __builtin_trap(); // behavior undefined if virtual wave size exceeds hardware limit
     }
 
     /// \brief Performs reduction across threads in a logical warp.
@@ -277,10 +272,7 @@ public:
         {
             if(VirtualWaveSize > ::rocprim::arch::wavefront::size())
             {
-                ROCPRIM_PRINT_ERROR_ONCE(
-                    "Specified warp size exceeds current hardware supported warp "
-                    "size. Aborting warp sort.");
-                return;
+                __builtin_trap(); // behavior undefined if virtual wave size exceeds hardware limit
             }
         }
         base_type::reduce(input, output, valid_items, storage, reduce_op);
@@ -295,9 +287,7 @@ public:
         typename std::enable_if<(FunctionWarpSize > arch::wavefront::max_size()), void>::type
     {
         (void)reduce_op;
-        ROCPRIM_PRINT_ERROR_ONCE("Specified warp size exceeds current hardware supported warp "
-                                 "size. Aborting warp sort.");
-        return;
+        __builtin_trap(); // behavior undefined if virtual wave size exceeds hardware limit
     }
 
     /// \brief Performs head-segmented reduction across threads in a logical warp.
@@ -333,10 +323,7 @@ public:
         {
             if(VirtualWaveSize > ::rocprim::arch::wavefront::size())
             {
-                ROCPRIM_PRINT_ERROR_ONCE(
-                    "Specified warp size exceeds current hardware supported warp "
-                    "size. Aborting warp sort.");
-                return;
+                __builtin_trap(); // behavior undefined if virtual wave size exceeds hardware limit
             }
         }
         base_type::head_segmented_reduce(input, output, flag, storage, reduce_op);
@@ -353,9 +340,7 @@ public:
         typename std::enable_if<(FunctionWarpSize > arch::wavefront::max_size()), void>::type
     {
         (void)reduce_op;
-        ROCPRIM_PRINT_ERROR_ONCE("Specified warp size exceeds current hardware supported warp "
-                                 "size. Aborting warp sort.");
-        return;
+        __builtin_trap(); // behavior undefined if virtual wave size exceeds hardware limit
     }
 
     /// \brief Performs tail-segmented reduction across threads in a logical warp.
@@ -391,10 +376,7 @@ public:
         {
             if(VirtualWaveSize > ::rocprim::arch::wavefront::size())
             {
-                ROCPRIM_PRINT_ERROR_ONCE(
-                    "Specified warp size exceeds current hardware supported warp "
-                    "size. Aborting warp sort.");
-                return;
+                __builtin_trap(); // behavior undefined if virtual wave size exceeds hardware limit
             }
         }
         base_type::tail_segmented_reduce(input, output, flag, storage, reduce_op);
@@ -411,9 +393,7 @@ public:
         typename std::enable_if<(FunctionWarpSize > arch::wavefront::max_size()), void>::type
     {
         (void)reduce_op;
-        ROCPRIM_PRINT_ERROR_ONCE("Specified warp size exceeds current hardware supported warp "
-                                 "size. Aborting warp sort.");
-        return;
+        __builtin_trap(); // behavior undefined if virtual wave size exceeds hardware limit
     }
 };