+++ CHANGELOG.md
@@ -8,6 +8,14 @@ Full documentation for rocPRIM is available at [https://rocm.docs.amd.com/projec
 
 * Updated config system to pick better fallback configs for untuned GPUs.
 
+### Upcoming changes
+
+* Deprecated the `ROCPRIM_PRINT_ERROR_ONCE` macro.
+
+### Changed
+
+* Changed various APIs with undefined behaviors to abort with a trap instead of printing a runtime error with `ROCPRIM_PRINT_ERROR_ONCE` 
+
 ## rocPRIM 4.2.0 for ROCm 7.2
 
 ### Added